ref: separate logic on start into other component remove: useless console.log
